security: add authorization checks to all content management operations

+++ app/Livewire/Topic/Create.php
@@ -14,6 +14,11 @@ class Create extends Component
 
     public $editMode = false;
 
+    public function boot()
+    {
+        $this->authorize('create', Topic::class);
+    }
+
     public function updateEditMode(Topic $topic)
     {
         $this->topic = $topic;
@@ -66,6 +71,8 @@ class Create extends Component
 
     public function deleteTopic(Topic $topic)
     {
+        $this->authorize('delete', $topic);
+
         $topic->delete();
 
         $this->topic = null;
